refactor: 메인 페이지 메인 페이지 라우팅 개선 및 로그인/로그아웃 UI 추가
