+++ src/App.js
@@ -7,7 +7,6 @@ import MovieRecommendation from './components/MovieRecommendation';
 import MovieDetail from './components/MovieDetail';
 import Calendar from './components/Calendar';
 import Profile from './components/Profile';
-import HomeScreen from './components/HomeScreen';
 import SearchModal from './components/SearchModal';
 import { useAuth } from './hooks/useAuth';
 
@@ -74,8 +73,7 @@ function MovieDetailWrapper() {
 }
 
 function App() {
-  const { isLoading, isAuthenticated } = useAuth();
-  const [isGuest, setIsGuest] = useState(false);
+  const { isLoading } = useAuth();
 
   // 로딩 중일 때 표시
   if (isLoading) {
@@ -90,13 +88,8 @@ function App() {
   return (
     <Router>
       <Routes>
-        {/* 인증 X && 게스트 X -> 홈 스크린 */}
-        {!isAuthenticated && !isGuest ? (
-          <Route path="*" element={<HomeScreen onStart={() => setIsGuest(true)} />} />
-        ) : (
-          /* 인증된 사용자 또는 게스트는 메인 앱으로 */
-          <Route path="*" element={<AppLayout />} />
-        )}
+        {/* 모든 사용자는 바로 메인 앱으로 이동 */}
+        <Route path="*" element={<AppLayout />} />
       </Routes>
     </Router>
   );
