+++ src/components/MainContent.js
@@ -1,5 +1,7 @@
 import React, { useState, useEffect, useCallback, useRef } from 'react';
 import { useMovies } from '../hooks/useMovies';
+import { useAuth } from '../hooks/useAuth';
+import KakaoLogin from './KakaoLogin';
 import './MainContent.css';
 
 const MainContent = ({ onMovieClick }) => {
@@ -11,6 +13,28 @@ const MainContent = ({ onMovieClick }) => {
     loadMoreMovies,
     refreshMovies 
   } = useMovies();
+
+  // 인증 관련 상태
+  const { user, isAuthenticated, isLoading: authLoading, error: authError, login, logout, clearError } = useAuth();
+
+  // 로그인 핸들러
+  const handleLoginSuccess = async (kakaoAccessToken) => {
+    try {
+      clearError();
+      await login(kakaoAccessToken);
+      console.log('MainContent: 로그인 성공');
+    } catch (err) {
+      console.error('MainContent: 로그인 실패', err);
+    }
+  };
+
+  const handleLoginError = (errorMessage) => {
+    console.error("Kakao SDK 에러:", errorMessage);
+  };
+
+  const handleLogout = () => {
+    logout();
+  };
 
   // 캐러셀을 위한 상태
   const [currentCarouselIndex, setCurrentCarouselIndex] = useState(0);
@@ -143,6 +167,31 @@ const MainContent = ({ onMovieClick }) => {
 
   return (
     <main className="main-content">
+      {/* Compact User Auth Section - 우측 상단 */}
+      <div className="compact-auth-section">
+        {isAuthenticated ? (
+          <div className="compact-user-info">
+            <span className="compact-welcome">안녕하세요, {user?.name || '사용자'}님!</span>
+            <button className="compact-logout-button" onClick={handleLogout}>
+              로그아웃
+            </button>
+          </div>
+        ) : (
+          <div className="compact-login-section">
+            <KakaoLogin 
+              onLoginSuccess={handleLoginSuccess} 
+              onLoginError={handleLoginError}
+            />
+            {authError && (
+              <div className="compact-auth-error">
+                <span>{authError}</span>
+                <button onClick={clearError}>×</button>
+              </div>
+            )}
+          </div>
+        )}
+      </div>
+
       {/* Carousel Section */}
       <section 
         className="carousel-section"
